requirements.txt
cloudinary==1.8.0
Django==1.10
django-pagedown==0.1.3
virtualenv==15.1.0
virtualenv-clone==0.2.6
